update comments and version #
the output from C-c C-c now essentially works with compilation-mode, so
you can use `next-error' and `previous-error' to navigate.

also, I've included commands to run gnuplot on the buffer and region.
these do not require the buffer to be visiting a file (or for the file
to be saved), but they don't produce output compatible with
compilation-mode.